Add project scaffolding
This includes

.gitignore: ignores common python build/generated files python projects

CONTRIBUTING.rst: with instructions for how to contribute. This file's
                  presence is required to comply with Google's open
                  source policy

README.rst: Provides a high-level description of the project. This
            file's presence and is in line Google's open source policy,
            and

{test-,}requirements.txt: declare package dependencies; is used to setup
                          virtual environment during development

tox.ini: configuration for tox a python standard tool for automating
         test, lint and doc tasks

{test-,}requirements.txt: declare dependencies, used to setup virtual
                          environments

Also, the following python packages are started

google/apigen: contains message and client classes
               generated using google apitools

google/scc: this is the package that will contain the service control
            client code.

Code-Review notes:
 - fixed python versions in setup.py
 - corrected copyright year

Change-Id: Iff11c40559f9f69266cc4415b0533c155c2bbcbf