Copy the GAX path template parser
Parsing the methods in the service configuration requires interpretation
of http temmplates in the methods http rules.

gax-python has a class for doing this, but it would be inappropriate
to take a dependency on that package to use it.

Instead, google.scc takes a copy of the class in this change

Change-Id: Iccbc5713094003c16d4c637a9b52b03c4ded7f1f