Adds the client and wsgi integraton layer.
- google.scc.client.Client is standalone client that performs service
control. It has #start and
flushing of report and check requests to the underlying http transport

- google.scc.wsgi.Middleware that uses a Client to provide a piece
of WSGI middleware that performs service control.

- TODO: unittests and examples, but most nearly all the major pieces of
behaviour are present.  Test coverage is now at 87%, before this PR it
was at 97%, it'd be nice to get back there given that this python.
However, there is a lot in this CL, so unittests will come in subsequent
CRs.

Change-Id: I7ffafb258c7e3e57b87f1b4a1d6e6c49eeecef0b