remove namespace_packages option from setup.py (#10)
* remove namespace_packages option from setup.py

We need __init__.py to be installed during the package installation
process, because "google" is a namespace package that we have to share
with the App Engine google.net.proto library. If we specify this option,
the files are omitted, which breaks Endpoints importing.

The setuptools help[1] says it should be declared, but it doesn't really
state that it's required. I've been unable to determine a solution that
works without removing this option. One suggested solution was to add
the vendor directory as a site package to properly process .pth
files. However, this does not seem to address namespace packages.

[1]: http://setuptools.readthedocs.io/en/latest/setuptools.html#namespace-packages

* bump version to 1.0.0b8