regenerate generated API surface
This is a non-trivial change, due to the previous hand-editing of the
generated surface. Previously, the Service Management messages were
included in the servicecontrol_v1_messages module. Now, to make it
easier to maintain, these messages are split into
servicemanagement_v1_messages. Previously, all messages were imported by
importing servicecontrol_v1_messages as messages. The import is now
either servicecontrol_v1_messages as sc_messages or
servicemanagement_v1_messages as sm_messages.

This change also reduces the catch-all exception in the client module
to only catch apitools errors. This may be expanded later, but this was
overly broad and masking errors in the unit test. The generated service
has changed methods to be upper case, and even with the code still
using the old method names, the tests were passing.