re-add per-consumer metric reporting
This change is a partial, manual reversion of #12, re-adding
per-consumer metric reporting. In addition to this, per-consumer
metrics are now only reported when service control returns a consumer
project number to report on. If it does, per-consumer metrics, as well
as a consumer project label are included in the report.