Retry service config fetch if it fails. (#80)
* Comply more thoroughly with WSGI spec

* Retry service config fetch if it fails.

add_all() used to directly fetch the service config and then either
construct the middleware or replace the app with a 503-serving
app. Now it will return an instance of ConfigFetchWrapper, which acts
as a WSGI middleware. If the service config has been loaded, it just
passes requests into the ordinary middleware and thence to the app.

But if the service config can't be loaded, it will retry. If threading
is available, it spawns a background thread to retry with exponential
backoff; if threading isn't available, it retries on every request.

Fixes #76.